requirements.txt
certbot>=2.9.0
setuptools
requests
mock
requests-mock
sh-nic-api>=1.0.3
zope.interface

build
